chg(login): consolidate /login /signup /welcome
